Add useCallback wrapper
Since `getQuote` is defined within the scope of the component, if I understand correctly, it will be recomputed on every render. We don't want this for performance, so I wrapped it in `useCallback`, passing `[]` as the second argument in order to tell React to not recompute the function on every render.